moves project into a pnpm monorepo
