+++ .github/workflows/deploy.yml
@@ -32,14 +32,14 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-      - name: npm setup
-        uses: actions/setup-node@v4
+      - uses: pnpm/action-setup@v2
         with:
-          node-version: 20
-      - name: Install
-        run: npm install
+          version: 8
+          run_install: |
+            - recursive: true
+              args: [--frozen-lockfile, --strict-peer-dependencies]
       - name: Build
-        run: npm run build
+        run: pnpm build
         env:
           BASE_URL: ${{ github.event.repository.name }}
       - name: Setup Pages
