+++ README.md
@@ -1,0 +1,12 @@
+## Local Setup
+
+### Prerequisites
+
+- Node.js: https://nodejs.org/en/
+- pnpm: https://pnpm.io/installation
+
+### Install packages
+
+```sh
+pnpm i
+```
