+++ apps/web/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "web",
+  "private": true,
+  "version": "0.0.0",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "@libs/webgl": "workspace:*"
+  }
+}
