+++ apps/web/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "include": ["src"],
+  "references": [{ "path": "./tsconfig.node.json" }]
+}
