+++ libs/webgl/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@libs/webgl",
+  "private": true,
+  "version": "0.0.0",
+  "main": "./src/index.ts"
+}
