+++ libs/webgl/src/programs/shaders/color.fragment.glsl
