+++ libs/webgl/src/types/arrays.d.ts
