+++ libs/webgl/src/types/declarations.d.ts
