+++ libs/webgl/src/types/vite-env.d.ts
