+++ libs/webgl/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "include": ["src"]
+}
